Implemented Partial DOM Tree building based on Sel Targets
Elements are not added/linked as children in the built DOM tree, unless
the streamer is explicitely told to keep them.
Selectors automatically have the children of their targets linked/kept,
so no need to worry there.
Also, operations like Element.getChildren() and .getText(), which
trigger `doComplete()`, specify that the "completed" children should be
kept.